% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\name{ase}
\alias{ase}
\title{Spectral Embedding of Adjacency Matrices}
\usage{
ase(g, dim)
}
\arguments{
\item{g}{The input graph, directed or undirected.}

\item{dim}{An integer scalar. This value is the embedding dimension of the
spectral embedding. Should be smaller than the number of vertices. The
largest \code{no}-dimensional non-zero singular values are used for the
spectral embedding.}
}
\value{
A list containing with entries: \item{X}{Estimated latent positions,
an \code{n} times \code{no} matrix, \code{n} is the number of vertices.}
\item{Y}{\code{NULL} for undirected graphs, the second half of the latent
positions for directed graphs, an \code{n} times \code{no} matrix, \code{n}
is the number of vertices.} \item{D}{The eigenvalues (for undirected graphs)
or the singular values (for directed graphs) calculated by the algorithm.}
\item{options}{A named list, information about the underlying ARPACK
computation. See \code{\link{arpack}} for the details.}
}
\description{
Spectral decomposition of the adjacency matrices of graphs.
}
\details{
This function computes a \code{no}-dimensional Euclidean representation of
the graph based on its adjacency matrix, \eqn{A}. This representation is
computed via the singular value decomposition of the adjacency matrix,
\eqn{A=UDV^T}.In the case, where the graph is a random dot product graph
generated using latent position vectors in \eqn{R^{no}} for each vertex, the
embedding will provide an estimate of these latent vectors.

For undirected graphs the latent positions are calculated as
\eqn{X=U^{no}D^{1/2}}{U[no] sqrt(D[no])}, where \eqn{U^{no}}{U[no]} equals
to the first \code{no} columns of \eqn{U}, and \eqn{D^{1/2}}{sqrt(D[no])} is
a diagonal matrix containing the top \code{no} singular values on the
diagonal.

For directed graphs the embedding is defined as the pair
\eqn{X=U^{no}D^{1/2}}{U[no] sqrt(D[no])} and \eqn{Y=V^{no}D^{1/2}}{V[no]
sqrt(D[no])}. (For undirected graphs \eqn{U=V}, so it is enough to keep one
of them.)
}
\examples{

## A small graph
lpvs <- matrix(rnorm(200), 20, 10)
lpvs <- apply(lpvs, 2, function(x) { return (abs(x)/sqrt(sum(x^2))) })
RDP <- sample_dot_product(lpvs)
embed <- embed_adjacency_matrix(RDP, 5)
}
\references{
Sussman, D.L., Tang, M., Fishkind, D.E., Priebe, C.E.  A
Consistent Adjacency Spectral Embedding for Stochastic Blockmodel Graphs,
\emph{Journal of the American Statistical Association}, Vol. 107(499), 2012
}
\seealso{
\code{\link{sample_dot_product}}
}
\keyword{graphs}
