% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celegans.male.chem}
\alias{celegans.male.chem}
\title{C. Elegans Male Chemical Connectome}
\format{An \code{\link[igraph]{igraph}} object containing the chemical graph:
\describe{
\item{\code{V(celegans.male.chem)$name}}{The name of the C. Elegans neuron.}
\item{\code{V(celegans.male.chem)$type}}{The type of the C. Elegans neuron. Either \code{M} motor neuron, \code{I} inter neuron, or \code{S} sensory neuron.}
\item{\code{V(celegans.male.chem)$hemisphere}}{The hemisphere the C. Elegans neuron is found in. Either \code{R} right, \code{L} left, or \code{B} bilateral, ring neuron spanning both hemispheres, etc.}
}}
\usage{
celegans.male.chem
}
\description{
A graph consisting of a synapse-level map of the chemical connections in the brain of a male Caenorhabditis (C.) Elegans.
}
\references{
Cook et al. The complete connectome of Caenorhabditis Elegans. 2018.
}
\keyword{datasets}
