% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly.mushroom.left}
\alias{fly.mushroom.left}
\title{Drosophila Left Mushroom Body Connectome}
\format{An \code{\link[igraph]{igraph}} object containing the graph:
\describe{
\item{\code{V(fly.mushroom.left)$name}}{The name of the Drosophila neuron.}
\item{\code{V(fly.mushroom.left)$type}}{The type of the Drosophila neuron. Either \code{K} Kenyon Cell, \code{I} MBIN cell line,  \code{O} MBON cell ine, or \code{P} PN cell line.}
\item{\code{V(fly.mushroom.left)$claw}}{The claw each Kenyon Cell is part of in the mushroom body. Cells not marked with \code{type} as \code{K} are not assigned a value.}
\item{\code{V(fly.mushroom.left)$dist}}{The distance along the claw each Kenyon Cell is part of in the mushroom body. Cells not marked with \code{type} as \code{K} are not assigned a value.}
}}
\usage{
fly.mushroom.left
}
\description{
A binary graph consisting of a synapse-level map of the connections in the left mushroom body of a Drosophila (fruit fly).
}
\references{
Eichler, K. et al. The complete wiring diagram of a high-order learning and memory center, the insect mushroom body.
\emph{Nature}, 2017.
}
\keyword{datasets}
