% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclust.R
\name{gclust}
\alias{gclust}
\title{Select number of mixture components for Gaussian Mixture Model via maximizing BIC.}
\usage{
gclust(X, K = 2)
}
\arguments{
\item{X}{A matrix object with n rows (data points) by d columns (dimensions).}

\item{K=2}{The maximum number of components to be considered. Model selected from 1 to K components.}
}
\value{
The number of components that maximizes BIC.
}
\description{
\code{glust} selects optimal number of mixture components to describe data in GMM.
}
