% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gs.as_adj}
\alias{gs.as_adj}
\title{From Graph Helper}
\usage{
gs.as_adj(g, edge.attr = NULL)
}
\arguments{
\item{g}{an igraph object or an adjacency matrix.}

\item{edge.attr}{if g is an igraph object, the name of the edge attribute to use for weights.}
}
\value{
An adjacency matrix corresponding to the graph.
}
\description{
function to standardize format internally for usage with n by n adjacency matrices
}
\author{
Eric Bridgeford
}
\keyword{internal}
