% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimselect.R
\name{gs.dim.select}
\alias{gs.dim.select}
\title{Dimensionality selection for singular values using profile likelihood.}
\usage{
gs.dim.select(X, k = NULL, n = 3, threshold = FALSE, plot = FALSE)
}
\arguments{
\item{X}{an object of class \link[igraph]{graph}, a numeric/complex matrix or 2-D array with \code{n} rows and \code{d} columns,
or a one-dimensional vector of class \code{numeric} containing ordered singular values.
If \code{X} is a \link[igraph]{graph}, the graph is embedded into \code{k} dimensions with \link[igraph]{embed_adjacency_matrix}. If
\code{X} is a \code{matrix} or 2-D \code{array}, the matrix is embedded into \code{k} dimensions with \link[base]{svd}.
\link[igraph]{embed_adjacency_matrix} if \code{X} is a \link[igraph]{graph} to obtain the singular values.}

\item{k}{If \code{X} is of class \link[igraph]{graph} or a numeric/complex matrix or 2-D array, an integer scalar indicating
the embedding dimensionality of the spectral embedding. Should have \code{k < length(V(X))} if \code{X} is of class
\link[igraph]{graph}, or \code{k < min(dim(X))} if \code{X} is a numeric/complex matrix or 2-D array.}

\item{n}{default value: 3; the number of returned elbows.}

\item{threshold}{either \code{FALSE} or an object of class \code{numeric}. If threshold is of class \code{numeric}, then all
the elements that are not larger than the threshold will be ignored.}

\item{plot}{logical. When \code{TRUE}, the return object includes a plot depicting the elbows.}
}
\value{
list containing the following:
\item{\code{value}}{The singular values associated with each elbow in \code{elbow}.}
\item{\code{elbow}}{The indices of the elbows.}
\item{\code{plot}}{If \code{plot} is \code{TRUE}, contains a scree plot annotated with the elbows.}
}
\description{
Select the number of significant singular values, by finding the
\sQuote{elbow} of the scree plot, in a principled way.
}
\details{
The input of the function is a numeric vector which contains the measure of
\sQuote{importance} for each dimension.

For spectral embedding, these are the singular values of the adjacency
matrix. The singular values are assumed to be generated from a Gaussian
mixture distribution with two components that have different means and same
variance. The dimensionality \eqn{d} is chosen to maximize the likelihood
when the \eqn{d} largest singular values are assigned to one component of
the mixture and the rest of the singular values assigned to the other
component.

This function can also be used for the general separation problem, where we
assume that the left and the right of the vector are coming from two Normal
distributions, with different means, and we want to know their border. See
examples below.
}
\examples{

# Generate the two groups of singular values with
# Gaussian mixture of two components that have different means
sing.vals  <- c( rnorm (10, mean=1, sd=1), rnorm(10, mean=3, sd=1) )
dim.chosen <- dimselect(sing.vals)
dim.chosen

# Sample random vectors with multivariate normal distribution
# and normalize to unit length
lpvs <- matrix(rnorm(200), 10, 20)
lpvs <- apply(lpvs, 2, function(x) { (abs(x) / sqrt(sum(x^2))) })
RDP.graph  <- sample_dot_product(lpvs)
gs.dim.select( embed_adjacency_matrix(RDP.graph, 10)$D )

# Sample random vectors with the Dirichlet distribution
lpvs.dir    <- sample_dirichlet(n=20, rep(1, 10))
RDP.graph.2 <- sample_dot_product(lpvs.dir)
gs.dim.select( embed_adjacency_matrix(RDP.graph.2, 10)$D )

# Sample random vectors from hypersphere with radius 1.
lpvs.sph    <- sample_sphere_surface(dim=10, n=20, radius=1)
RDP.graph.3 <- sample_dot_product(lpvs.sph)
gs.dim.select( embed_adjacency_matrix(RDP.graph.3, 10)$D )

}
\references{
M. Zhu, and A. Ghodsi (2006). Automatic dimensionality selection
from the scree plot via the use of profile likelihood. \emph{Computational
Statistics and Data Analysis}, Vol. 51, 918--930.
}
\seealso{
\code{\link{embed_adjacency_matrix}}
}
\author{
Youngser Park \email{youngser@jhu.edu}, Gabor Csardi \email{csardi.gabor@gmail.com}, and Eric Bridgeford \email{ericwb95@gmail.com}.
}
\keyword{graphs}
