% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_xfms.R
\name{gs.edge.log}
\alias{gs.edge.log}
\title{Run pass-to-log on a weighted graph.}
\usage{
gs.edge.log(g, edge.attr = NULL, base = exp(1), eps = 1,
  output.type = "matrix")
}
\arguments{
\item{g}{an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices.}

\item{edge.attr}{if \code{g} is a \code{\link[igraph]{igraph}}, the name of the attribute to use for weights. Defaults to \code{NULL}, which assumes the graph is binary.
\describe{
\item{\code{is.null(edge.attr)}}{constructs sbm on the graph as a binary adjacency matrix.}
\item{\code{is.character(edge.attr)}}{constructs sbm of the graph the graph as a weighted adjacency matrix, with edge-weights for \code{E(g)} given by \code{get.edge.attribute(g, attr=edge.attr)}.}
}}

\item{base}{the base for the logarithm operation. Defaults to base \code{exp(1)}, or the natural logarithm. Options are real numbers greater than \code{0}.}

\item{eps}{the offset for taking the logarithm, in the event that there are entries of \code{g} that are less than or equal to \code{0}.}

\item{output.type}{the type of output to produce for the between community expectations. Defaults to \code{matrix}.
\describe{
\item{\code{"matrix"}}{produces a matrix for the between-communitity interactions.}
\item{\code{"graph"}}{produces an \code{\link[igraph]{igraph}} object for the between-community interactions.}
}}
}
\value{
an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices depending on \code{output.type}.
}
\description{
Extracts edge weight vector \eqn{W} from a graph and replaces it with \eqn{log(W + eps)} where \eqn{log} is the logarithm operation.
}
\author{
Eric Bridgeford \href{mailto:ericwb95@gmail.com}{ericwb95@gmail.com}
}
