% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_xfms.R
\name{gs.edge.ptr}
\alias{gs.edge.ptr}
\title{Run pass-to-rank on a weighted graph.}
\usage{
gs.edge.ptr(g, edge.attr = NULL, output.type = "matrix")
}
\arguments{
\item{g}{an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices.}

\item{edge.attr}{if \code{g} is a \code{\link[igraph]{igraph}}, the name of the attribute to use for weights. Defaults to \code{NULL}, which assumes the graph is binary.
\describe{
\item{\code{is.null(edge.attr)}}{constructs sbm on the graph as a binary adjacency matrix.}
\item{\code{is.character(edge.attr)}}{constructs sbm of the graph the graph as a weighted adjacency matrix, with edge-weights for \code{E(g)} given by \code{get.edge.attribute(g, attr=edge.attr)}.}
}}

\item{output.type}{the type of output to produce for the between community expectations. Defaults to \code{matrix}.
\describe{
\item{\code{"matrix"}}{produces a matrix for the between-communitity interactions.}
\item{\code{"graph"}}{produces an \code{\link[igraph]{igraph}} object for the between-community interactions.}
}}
}
\value{
an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices depending on \code{output.type}.
}
\description{
It extracts a (non-zero) edge weight vector \eqn{W} from a graph and replaces it with \eqn{2*R / (|E|+1)} where \eqn{R} is the rank of \eqn{W} and \eqn{|E|} is the number of edges. This does 'no-op' for an unweighted or binary graph.
}
\author{
Eric Bridgeford \href{mailto:ericwb95@gmail.com}{ericwb95@gmail.com}
}
