% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_xfms.R
\name{gs.edge.threshold}
\alias{gs.edge.threshold}
\title{Threshold a weighted graph.}
\usage{
gs.edge.threshold(g, edge.attr = NULL, threshold = 0, method = "absolute",
  output.type = "matrix")
}
\arguments{
\item{g}{an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices.}

\item{edge.attr}{if \code{g} is a \code{\link[igraph]{igraph}}, the name of the attribute to use for weights. Defaults to \code{NULL}, which assumes the graph is binary.
\describe{
\item{\code{is.null(edge.attr)}}{constructs sbm on the graph as a binary adjacency matrix.}
\item{\code{is.character(edge.attr)}}{constructs sbm of the graph the graph as a weighted adjacency matrix, with edge-weights for \code{E(g)} given by \code{get.edge.attribute(g, attr=edge.attr)}.}
}}

\item{threshold}{the threshold for removing edges. Defaults to 0. Sets edges thresholded away to \code{0} and edges included in the threshold range to \code{1}.}

\item{method}{The method to use for thresholding. Defaults to \code{"absolute"}.
\itemize{
\item{"absolute"}{Remove edges less than \code{threshold}.}
\item{"percentile"}{Remove edges less than the \code{threshold} percentile. If \code{method} is \code{"percentile"}, \code{threshold} should be less than 1.`}
}}

\item{output.type}{the type of output to produce for the between community expectations. Defaults to \code{matrix}.
\describe{
\item{\code{"matrix"}}{produces a matrix for the between-communitity interactions.}
\item{\code{"graph"}}{produces an \code{\link[igraph]{igraph}} object for the between-community interactions.}
}}

\item{eps}{the offset for taking the logarithm, in the event that there are entries of \code{g} that are less than or equal to \code{0}.}
}
\value{
an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices depending on \code{output.type}.
}
\description{
Extracts (non-zero) edge weight vector \eqn{W} from a graph and replaces it with \eqn{log(W + eps)} where \eqn{log} is the logarithm operation.
}
\author{
Eric Bridgeford \href{mailto:ericwb95@gmail.com}{ericwb95@gmail.com}
}
