% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{gs.embed.ase}
\alias{gs.embed.ase}
\title{Spectral Embedding of Adjacency Matrices}
\usage{
gs.embed.ase(g, k = NULL, edge.attr = NULL)
}
\arguments{
\item{g}{an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices.}

\item{edge.attr}{if \code{g} is a \code{\link[igraph]{igraph}}, the name of the attribute to use for weights. Defaults to \code{NULL}, which assumes the graph is binary.
\itemize{
\item{\code{is.null(edge.attr)} assumes unweighted.}
\item{\code{is.character(edge.attr)} assumes weighted with weights given by \code{edge.attr}.}
}}
}
\value{
A list containing the following:
\item{\code{X}}{an \code{n} by \code{k} matrix indicating the estimated latent positions, where \code{n} is the number of vertices of \code{g}.}
\item{\code{Y}}{\code{NULL} if \code{g} is undirected. If \code{g} is directed, \code{Y} is a \code{n} by \code{k} matrix indicating the second half of the latent positions.}
\item{D}{The eigenvalues (for undirected graphs) or the singular values (for directed graphs) associated with the latent positions.}
}
\description{
Spectral decomposition of the adjacency matrices of graphs.
}
\details{
This function computes a \code{k}-dimensional Euclidean representation of
the graph based on its adjacency matrix, \eqn{A}. This representation is
computed via the singular value decomposition of the adjacency matrix,
\eqn{A=UDV^T}.
}
\examples{

## A small graph
lpvs <- matrix(rnorm(200), 20, 10)
lpvs <- apply(lpvs, 2, function(x) { return (abs(x)/sqrt(sum(x^2))) })
RDP <- sample_dot_product(lpvs)
embed <- embed_adjacency_matrix(RDP, 5)
}
\references{
Sussman, D.L., Tang, M., Fishkind, D.E., Priebe, C.E.  A
Consistent Adjacency Spectral Embedding for Stochastic Blockmodel Graphs,
\emph{Journal of the American Statistical Association}, Vol. 107(499), 2012
}
\author{
Eric Bridgeford \href{mailto:ericwb95@gmail.com}{ericwb95@gmail.com}
}
\keyword{graphs}
