% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lse.R
\name{gs.embed.lse}
\alias{gs.embed.lse}
\title{Spectral Embedding of the Laplacian of a Graph}
\usage{
gs.embed.lse(g, k)
}
\arguments{
\item{g}{The input graph, directed or undirected.}

\item{k}{An integer scalar. Should be the case that \code{k < gorder(g)}. The
largest \code{k}-dimensions are retained from the spectral embedding.}
}
\value{
A list containing the following:
\item{\code{X}}{an \code{n} by \code{k} matrix indicating the estimated latent positions, where \code{n} is the number of vertices of \code{g}.}
\item{\code{Y}}{\code{NULL} if \code{g} is undirected. If \code{g} is directed, \code{Y} is a \code{n} by \code{k} matrix indicating the second half of the latent positions.}
\item{D}{The eigenvalues (for undirected graphs) or the singular values (for directed graphs) associated with the latent positions.}
\item{options}{A named list, information about the underlying ARPACK computation. See \code{\link{arpack}} for the details.}
}
\description{
Spectral decomposition of Laplacian matrices of graphs.
}
\details{
This function computes a \code{k}-dimensional Euclidean representation of
the graph based on its Laplacian matrix, \eqn{L}. This representation is
computed via the singular value decomposition of the Laplacian matrix.
}
\examples{

# A small graph
lpvs <- matrix(rnorm(200), 20, 10)
lpvs <- apply(lpvs, 2, function(x) { return (abs(x)/sqrt(sum(x^2))) })
RDP <- sample_dot_product(lpvs)
embed <- embed_laplacian_matrix(RDP, 5)
}
\references{
Sussman, D.L., Tang, M., Fishkind, D.E., Priebe, C.E.  A
Consistent Adjacency Spectral Embedding for Stochastic Blockmodel Graphs,
\emph{Journal of the American Statistical Association}, Vol. 107(499), 2012
}
\seealso{
\code{\link{embed_adjacency_matrix}}, \code{\link{sample_dot_product}}
}
\author{
Gabor Csardi \email{csardi.gabor@gmail.com}
}
\keyword{graphs}
