% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{gs.embed}
\alias{gs.embed}
\title{Embedding Helper}
\usage{
gs.embed(A, k = NULL)
}
\arguments{
\item{A}{the adjacency matrix to embed.}

\item{k}{the number of embedding dimensions.}
}
\description{
A function to help with embedding points.
}
\author{
Eric Bridgeford \href{mailto:ericwb95@gmail.com}{ericwb95@gmail.com}
}
\keyword{internal}
