% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni.R
\name{gs.omni}
\alias{gs.omni}
\title{Omnibus}
\usage{
gs.omni(A1, A2)
}
\arguments{
\item{A1}{the first \code{[n, n]} matrix.}

\item{A2}{the second \code{[n, n]} matrix.}
}
\value{
the omnibus matrix, as an object of class \code{\link[igraph]{igraph}}. The vertices are concatenated to form one
}
\description{
A function for creating an omnibus graph from arbitrarily many input graphs with
matched vertex sets.
}
\details{
Given \eqn{A_1, A_2, ..., A_m} a collection of (possibly weighted) adjacency matrices of a collection
of \eqn{m} undirected graphs with matched vertices. Then the \eqn{mn} by \eqn{mn} Omnibus Matrix has the
subgraph where \eqn{M_{ij} = 1/2 (A_i + A_j)}.
}
\author{
Youngser Park
}
