% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gs.output}
\alias{gs.output}
\title{Output Helper}
\usage{
gs.output(A, gref = NULL, edge.attr = NULL, output.type = "matrix")
}
\arguments{
\item{A}{the n x n adjacency matrix.}

\item{gref}{a reference graph to use for adding vertex attributes.}

\item{edge.attr}{the edge attribute to (optionally) add weights from the adjacency matrix to.}

\item{output.type}{"matrix" or "graph".}
}
\description{
Produces an output with the desired type.
}
\author{
Eric Bridgeford
}
\keyword{internal}
