% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gs.plot.heatmap_overlay}
\alias{gs.plot.heatmap_overlay}
\title{Graph Grid plot}
\usage{
gs.plot.heatmap_overlay(g, title = "", src.label = "Vertex",
  tgt.label = "Vertex", edge.attr = NULL, font.size = NULL,
  vertex.label = FALSE, vertex.attr = FALSE, edge.xfm = FALSE,
  eps = 1e-04, degree = FALSE)
}
\arguments{
\item{g}{input graph, as an igraph object. See \code{\link[igraph]{graph}} for details.}

\item{title}{the title for the square plot. Defaults to \code{""}.}

\item{src.label}{the source label for the graph. Defaults to \code{"Vertex"}.}

\item{tgt.label}{the target label for the graph. Defaults to \code{"Vertex"}.}

\item{edge.attr}{the name of the attribute to use for weights. Defaults to \code{NULL}. Can be a list of \code{edge.attr} if you want to overlay different edge-attributes on the same plot. Supports up to 4 edge-attributes at once.
\itemize{
\item{\code{is.null(edge.attr)} plots the graph as a binary adjacency matrix.}
\item{\code{is.character(edge.attr)} plot the graph as a weighted adjacency matrix, with edge-weights for \code{E(g)} given by \code{E(g)[[edge.attr]]}.}
}}

\item{font.size}{the default font size for the plot text. Axis/legend text is \code{font.size - 2}. Defaults to \code{NULL}.
\itemize{
\item{\code{is.null(font.size)} uses the default sizing for all fonts.}
\item{\code{!is.null(font.size)} uses \code{font.size} as the font sizing for the plots.}
}}

\item{vertex.label}{an attribute for naming the vertices. Defaults to \code{NULL}.
\itemize{
\item{\code{vertex.label==FALSE} name the vertices \code{V(g)} sequentially, as 1, 2, ... n.}
\item{\code{vertex.label==TRUE} name the vertices \code{V(g)} as \code{V(g)$name}.}
}}

\item{vertex.attr}{an attribute to color vertices. Defaults to \code{FALSE}.
\itemize{
\item{\code{vertex.attr==FALSE} assumes no grouping of the vertices, and adds no color accordingly.}
\item{\code{is.character(vertex.attr)} assumes a grouping of the vertices given for \code{V(g)} by \code{V(G)[[vertex.attr]]}, and groups the vertices in \code{V(g)} into ordered blocks with color-coding.}
}}

\item{edge.xfm}{transform the edge-weights. Defaults to \code{FALSE}. Can be a list of \code{edge.xfm} if you want to overlay different edge-attributes on the same plot in the plot with different transforms for each.
\itemize{
\item{\code{edge.xfm==FALSE} do not transform the edge-weights.}
\item{\code{edge.xfm == "log"} transform the edge values, using the natural-logarithm operation. See \code{\link[base]{log}} for details. Does not work if there are negative edge-weights. Pads with \code{eps << min(edge-weight)} if there are entries of zero in the graph.}
\item{\code{edge.xfm == "log10"} transform the edge values, using the logarithm-base-10 operation. See \code{\link[base]{log10}} for details. Does not work if there are negative edge-weights. Pads with \code{eps << min(edge-weight)} if there are entries of zero in the graph.}
}}

\item{eps}{if you specify an \code{edge.xfm} that is logarithmic, indicate the padding that zero entries should receive. Defaults to \code{.0001}. \code{eps} should be \code{<< min(edge-weight)}.}

\item{degree}{Whether to plot the marginal vertex degrees. Defaults to \code{FALSE}.}
}
\value{
the graph/graphs as a plot.
}
\description{
A function that plots an igraph object, as a series of overlapping heatmaps, with intensity denoted by the colorscale of dots on the heatmap.
}
\author{
Eric Bridgeford
}
