% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gs.plot.matrix.pairs}
\alias{gs.plot.matrix.pairs}
\title{Matrix Pairs Plot}
\usage{
gs.plot.matrix.pairs(X, k = NULL, x.label = "", y.label = "",
  title = "", pt.label = "Point", font.size = NULL)
}
\arguments{
\item{X}{input matrix as a 2-d data frame. Should be \code{[n, d]} dimensions for \code{n} points and \code{[d]} dimensions. The points (rows) rows
of \code{X} points will be colored according to their \code{rownames} assignment.}

\item{k}{the maximum number of dimensions you want to plot. Defaults to \code{NULL}, which plots all \code{d} dimensions.}

\item{x.label}{the x label for the pairs plot. Defaults to \code{""}.}

\item{y.label}{the y label for the pairs plot. Defaults to \code{""}.}

\item{title}{the title for the plot. Defaults to \code{""}.}

\item{pt.label}{the name of the point labels. If the columns of \code{X} are named, will color the points}

\item{font.size}{the default font size for the plot text. Axis/legend text is \code{font.size - 2}. Defaults to \code{NULL}.
\itemize{
\item{\code{is.null(font.size)} uses the default sizing for all fonts.}
\item{\code{!is.null(font.size)} uses \code{font.size} as the font sizing for the plots.}
}}
}
\value{
the latent positions as a pairs plot.
}
\description{
A function that plots a matrix as a pairs plot.
}
\author{
Eric Bridgeford
}
