% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gs.plot.plot_adjacency}
\alias{gs.plot.plot_adjacency}
\title{Graph plot}
\usage{
gs.plot.plot_adjacency(g, title = "", src.label = "Vertex",
  tgt.label = "Vertex", edge.attr = FALSE, font.size = NULL,
  vertex.label = FALSE, vertex.attr = FALSE, edge.xfm = FALSE,
  eps = 1e-04)
}
\arguments{
\item{g}{input graph, as an igraph object. See \code{\link[igraph]{graph}} for details.}

\item{title}{the title for the square plot. Defaults to \code{""}.}

\item{src.label}{the source label for the graph. Defaults to \code{"Vertex"}.}

\item{tgt.label}{the target label for the graph. Defaults to \code{"Vertex"}.}

\item{edge.attr}{the name of the attribute to use for weights. Defaults to \code{NULL}.
\itemize{
\item{\code{is.null(edge.attr)} plots the graph as a binary adjacency matrix.}
\item{\code{is.character(edge.attr)} plot the graph as a weighted adjacency matrix, with edge-weights for \code{E(g)} given by \code{E(g)[[edge.attr]]}.}
}}

\item{font.size}{the default font size for the plot text. Axis/legend text is \code{font.size - 2}. Defaults to \code{NULL}.
\itemize{
\item{\code{is.null(font.size)} uses the default sizing for all fonts.}
\item{\code{!is.null(font.size)} uses \code{font.size} as the font sizing for the plots.}
}}

\item{vertex.label}{an attribute for naming the vertices. Defaults to \code{NULL}.
\itemize{
\item{\code{vertex.label==FALSE} name the vertices \code{V(g)} sequentially, as 1, 2, ... n.}
\item{\code{vertex.label==TRUE} name the vertices \code{V(g)} as \code{V(g)$name}.}
}}

\item{vertex.attr}{an attribute to color vertices. Defaults to \code{FALSE}.
\itemize{
\item{\code{vertex.attr==FALSE} assumes no grouping of the vertices, and adds no color accordingly.}
\item{\code{is.character(vertex.attr)} assumes a grouping of the vertices given for \code{V(g)} by \code{V(G)[[vertex.attr]]}, and groups the vertices in \code{V(g)} into ordered blocks with color-coding.}
}}
}
\value{
the matrix as a plot.
}
\description{
A function that plots an igraph object, as an adjacency matrix.
}
\author{
Eric Bridgeford
}
