% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{gs.poisson.fit.zip}
\alias{gs.poisson.fit.zip}
\title{Zero-Inflated Poisson Model}
\usage{
gs.poisson.fit.zip(g, edge.attr = NULL, output = "matrix")
}
\arguments{
\item{g}{an \code{\link[igraph]{igraph}} object or an \code{nxn} adjacency matrix with \code{n} vertices.}

\item{edge.attr}{if \code{g} is a \code{\link[igraph]{igraph}}, the name of the attribute to use for weights. Defaults to \code{NULL}, which assumes the graph is binary.}

\item{communities}{an \code{n} vector containing the community label for each of the \code{n} vertices in \code{g}.}

\item{edge.attr}{if \code{g} is a \code{\link[igraph]{igraph}}, the name of the attribute to use for weights. Defaults to \code{NULL}, which assumes the graph is binary.
\describe{
\item{\code{is.null(edge.attr)}}{constructs sbm on the graph as a binary adjacency matrix.}
\item{\code{is.character(edge.attr)}}{constructs sbm of the graph the graph as a weighted adjacency matrix, with edge-weights for \code{E(g)} given by \code{get.edge.attribute(g, attr=edge.attr)}.}
}}

\item{output.type}{the type of output to produce for the between community expectations. Defaults to \code{matrix}.
\describe{
\item{\code{"matrix"}}{produces a matrix for the between-communitity interactions.}
\item{\code{"graph"}}{produces an \code{\link[igraph]{igraph}} object for the between-community interactions.}
}}
}
\value{
An object of class \code{ZIP} containing the following:
\describe{
\item{\code{n.v}}{the number of vertices in \code{g}.}
\item{\code{ne.nz}}{the number of non-zero edges in \code{g}.}
\item{\code{lambda}}{the poisson rate parameter for non-zero edges in \code{g}; also, the average non-zero edge weight of \code{g}.}
}
}
\description{
A function for fitting the 3-parameter (number of vertices, number of non-zero edges, lambda poisson rate parameter) Zero-Inflated Poisson Model to a graph.
A poisson model is fitted to the non-zero edges.
}
\author{
Eric Bridgeford
}
