% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{gs.sbm.fit}
\alias{gs.sbm.fit}
\title{Stochastic Block Model}
\usage{
gs.sbm.fit(g, community.attribute)
}
\arguments{
\item{g}{an \code{igraph} object. See \code{\link[igraph]{graph}} for details.}
}
\value{
the \code{[n, n]} probability matrix for each block, as an \code{igraph} object.
Vertices will be the \code{names(communities)}. Each vertex in the resulting graph will have an attribute \code{community} indicating the vertices in \code{graph} that
comprise the particular community the vertex summarizes See \code{\link[igraph]{graph}} for details.
}
\description{
A function to fit a stochastic block model to a weighted or unweighted graph.
}
\examples{
library(graphstats)

# sample a graph from an SBM
g <- sample_sbm(n=20, block.sizes=c(12, 8), pref.matrix=cbind(c(0.8, 0.2), c(0.2, 0.8)))
communities <- list(a=V(g)[1:12], b=V(g)[13:20])  # define communities of vertices in g
sbm.g <- gs.sbm.fit(g, communities)  # fit SBM to graph
}
\author{
Eric Bridgeford
}
