% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.batch.perm}
\alias{gs.siem.batch.perm}
\title{SIEM for Batch Detection}
\usage{
gs.siem.batch.perm(models, Z, i = 1, j = 2, tstat = gs.siem.batch.tstat,
  nperm = 1000)
}
\arguments{
\item{models}{[[n]] a list of the models fit to each of the n samples.}

\item{Z}{[n] an array of the labels associated with each model.}

\item{i=1}{the first edge set to use in the comparison.}

\item{j=2}{the second edge set to use in the comparison.}

\item{tstat=gs.siem.batch.tstat.delta}{the test statistic to use. Should operate on a pair of models.}

\item{nperm=1000}{the number of permutations for testing.}
}
\value{
tstat.alt the test statistic given the current Z orderings.

tstat.null the test statistics of each permutation of the Z orderings.

P the estimator of E[tstat.null >= tstat.alt.]
}
\description{
A function that computes a test statistic associated with a particular arrangement of graphs
using pairs of edge community estimators estimated by the SIEM.
}
\seealso{
\code{\link{gs.siem.fit}}
}
\author{
Eric Bridgeford
}
