% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.batch.test}
\alias{gs.siem.batch.test}
\title{SIEM for Batch Detection}
\usage{
gs.siem.batch.test(models, i = 1, j = 2, tstat = gs.siem.batch.tstat)
}
\arguments{
\item{models}{[[n]] a list of the models fit to each of the n samples. Each element must contain the following:
\itemize{
\item{model}{a model fit by \code{\link{gs.siem.fit}}}
\item{Z}{the batch id the particular model is from.}
}}

\item{i=1}{the first edge set to use in the comparison.}

\item{j=2}{the second edge set to use in the comparison.}

\item{tstat=gs.siem.batch.tstat.delta}{the test statistic to use. Should operate on a pair of models.}
}
\value{
tstat the test statistic.
}
\description{
A function that computes a test statistic associated with a particular arrangement of graphs
using pairs of edge community estimators estimated by the SIEM.
}
\seealso{
\code{\link{gs.siem.fit}}
}
\author{
Eric Bridgeford
}
