% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.batch.tstat}
\alias{gs.siem.batch.tstat}
\title{Test Statistic for Batch Detection}
\usage{
gs.siem.batch.tstat(set1est, set2est, Z)
}
\arguments{
\item{set1est}{[n] the first set of estimators for each element in the population.}

\item{set2est}{[n] the second set of estimators for each element in the population.}

\item{Z}{[n] the parameter labels for the batch each element is from.}
}
\value{
tstat the test statistic.
}
\description{
A function that computes a test statistic associated with a set of estimators.
}
\author{
Eric Bridgeford
}
