% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.fit}
\alias{gs.siem.fit}
\title{Structured Independent Edge Model, Single Sample}
\usage{
gs.siem.fit(graph, C, alt = "greater", ...)
}
\arguments{
\item{graph}{\code{[v, v]} the binary graph with \code{v} vertices.}

\item{C}{\code{[[r]]} a list of \code{r} communities where each element corresponds to a community of edges.
Note that \code{union(C[[j]])} over all \code{j} communities should be \code{1:v},
and \code{intersect(C[[j]])} over all \code{j} communities should be empty.}

\item{alt}{the alternative hypothesis for each edge-community test. Defaults to \code{'greater'}.
\itemize{
\item{'greater'}{p corresponds to \code{H0: x1 <= x2}, \code{HA: x1 > x2}. \code{R(x1, x2) = x1 > x2}}
\item{'neq'}{p corresponds to \code{H0: x1 = x2}, \code{HA: x1 != x2}. \code{R(x1, x2) = x1 != x2}}
\item{'less'}{p corresponds to \code{H0: x1 >= x2}, \code{HA: x1 < x2}. \code{R(x1, x2) = x1 < x2}}
}}

\item{...}{trailing args}
}
\value{
An object of class \code{"SIEM"} containing the following:
\item{\code{pr}}{\code{[r]} a probability vector corresponding to the probability of an edge connecting each edge community.}
\item{\code{var}}{\code{[r]} the variance of the probabilities estimated between edge communities.}
\item{\code{dpr}}{\code{[r, r]} an array consisting of the paired differences in probability where dpr_{ij} = pr_i - pr_j.}
\item{\code{dvar}}{\code{[r, r]} an array consisting of the variance of the estimate of the paired differences in probability, where dvar_{ij} = var_i + var_j.}
\item{\code{pv}}{\code{[r, r]} pv_{ij} is the p-value of false rejection of H0 that !R(pr_i, pr_j) in favor of HA that R(pr_j, pr_i).}
}
\description{
An independent-edge generalization of the traditional Stochastic Block Model.
}
\author{
Eric Bridgeford
}
