% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.model.params}
\alias{gs.siem.model.params}
\title{SIEM Model Parameters}
\usage{
gs.siem.model.params(data)
}
\arguments{
\item{data}{an array containing the edge data for a particular edge in a single graph.}
}
\value{
var the variance associated with a particular edge community.
}
\description{
A function that computes the parameters per-edge for SIEM.
}
\author{
Eric Bridgeford
}
