% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.sample.test}
\alias{gs.siem.sample.test}
\title{Estimator Sample Test}
\usage{
gs.siem.sample.test(x1, x2, var1, var2, df = NULL, n1 = NULL, n2 = NULL,
  alt = "greater")
}
\arguments{
\item{x1}{expectation value of an estimator.}

\item{x2}{expectation value of a second estimator.}

\item{var1}{the variance of the first estimator.}

\item{var2}{the variance of the second estimator.}

\item{df=NULL}{the number of degrees of freedom involved. Ignored if n1 and n2 are provided.}

\item{n1=NULL}{the number of observations for the first estimator. If n1 and n2 are not provided, the number of degrees of freedom is set to df.}

\item{n2=NULL}{the number of observations for the second estimator. If n1 and n2 are not provided, the number of degrees of freedom is set to df.
\itemize{
\item{1}{x1 and x2 are from different samples}
\item{2}{x1 and x2 are from different samples}
}}

\item{alt='greater'}{the alternative hypothesis for each edge test.
\itemize{
\item{'greater'}{p corresponds to H0: x1 <= x2, HA: x1 > x2. R(x1, x2) = x1 > x2}
\item{'neq'}{p corresponds to H0: x1 = x2, HA: x1 != x2. R(x1, x2) = x1 != x2}
\item{'less'}{p corresponds to H0: x1 >= x2, HA: x1 < x2. R(x1, x2) = x1 < x2}
}}
}
\value{
p is the p-value of false rejection of H0 that !R(x1, x2) in favor of HA that R(x1, x2).
}
\description{
A function that computes the p value of falsely rejecting H0 that x_i !? x_j in favor of HA that x_j ? x_i for x_i, x_j from the same sample.
}
