% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.sample_test}
\alias{gs.siem.sample_test}
\title{Estimator Sample Test}
\usage{
gs.siem.sample_test(x1, x2, var1, var2, nsamp, alt = "greater")
}
\arguments{
\item{x1}{value of an estimator.}

\item{x2}{value of a second estimator.}

\item{var1}{the variance of the first estimator.}

\item{var2}{the variance of the second estimator.}

\item{nsamp}{the number of samples involved. If x1 and x2 are from the same sample, nsamp is 1.
\itemize{
\item{1}{x1 and x2 are from different samples}
\item{2}{x1 and x2 are from different samples}
}}

\item{alt='greater'}{the alternative hypothesis for each edge test.
\itemize{
\item{'greater'}{p corresponds to H0: x1 <= x2, HA: x1 > x2}
\item{'neq'}{p corresponds to H0: x1 = x2, HA: x1 != x2}
\item{'less'}{p corresponds to H0: x1 >= x2, HA: x1 < x2}
}}
}
\value{
p is the p-value of false rejection of H0 that x1 !? x2 in favor of HA that x1 ? x2.
}
\description{
A function that computes the p value of falsely rejecting H0 that x_i !? x_j in favor of HA that x_j ? x_i for x_i, x_j from the same sample.
}
