% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gs.sims.er}
\alias{gs.sims.er}
\title{Erdos-Renyi Graphs}
\usage{
gs.sims.er(n, v, priors = c(0.5, 0.5), p = c(0.5, 0.5))
}
\arguments{
\item{n}{the number of graphs to simulate.}

\item{v}{the number of vertices.}

\item{priors}{the composition of the graphs in each class as a length \code{K} vector. Defaults to \code{c(0.5, 0.5)}. Note that \code{K=length(priors)} is the number of classes of graphs, and \code{sum(priors)} will be normalized to \code{1}.}

\item{p}{the probability of an edge existing for each class, as a length \code{K} vector. Defaults to \code{c(0.5, 0.5)}. Note that \code{K=length(p)} is the number of classes of graphs, and \code{0 <= p <= 1}.}
}
\value{
An object of class \code{"Simulation"} containing the following:
\item{\code{graphs}}{\code{[[n]][v, v]} the \code{n} graphs with \code{v} vertices as a list of adjacency matrices.}
\item{\code{Y}}{\code{[n]} the class labels for each of the \code{n} graphs.}
}
\description{
A function to simulate graphs from the Erdos-Renyi Random Graph Model.
}
\author{
Eric Bridgeford
}
