% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gs.sims.latent_pos}
\alias{gs.sims.latent_pos}
\title{Sample Graph from Latent Position}
\usage{
gs.sims.latent_pos(X, Y = NULL)
}
\arguments{
\item{X}{the latent positions, as a \code{n} by \code{k} matrix for \code{n} vertices and \code{k} blocks.}

\item{Y}{the second half of the latent positions for directed graphs, as a \code{n} by \code{k} matrix for \code{n} vertices and \code{k} blocks. Defaults to \code{NULL}, which assumes that the sampled graph is symmetric.}
}
\value{
An \link[igraph]{graph} object with \code{n} vertices, with latent positions given by \code{X} if \code{G} is symmetric and \code{X}, \code{Y} if nonsymmetric.
}
\description{
A function to simulate a binary graph given latent positions.
}
\examples{
library(graphstats)
data <- gs.sims.latent_pos(array(rnorm(100), dim=c(50, 2)))  # simulate 100 graphs with 10 vertices from the default model
}
\author{
Eric Bridgeford
}
