% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gs.sims.siem}
\alias{gs.sims.siem}
\title{Structured Independent Edge Model Graphs}
\usage{
gs.sims.siem(n, v, priors = c(0.5, 0.5), C = list(list(seq(1, floor(v^2/2)),
  seq(floor(v^2/2) + 1, v^2)), list(seq(1, floor(v^2/2)), seq(floor(v^2/2) + 1,
  v^2))), P = list(c(0.5, 0.5), c(0.5, 0.5)))
}
\arguments{
\item{n}{the number of graphs.}

\item{v}{the number of vertices.}

\item{priors}{the composition of the graphs in each class as a length \code{K} vector. Defaults to \code{c(0.5, 0.5)}.
Note that \code{K=length(priors)} is the number of classes of graphs, and \code{sum(priors)} will be normalized to \code{1}.}

\item{C}{the community membership of each edge in each class as a length \code{K} list of \code{length(r)} lists, where \code{r} is the number of communities for class \code{i}.
Note that \code{K=length(C)} is the number of classes of graphs, \code{union(C[[i]][[j]])} over all \code{j} communities for class \code{i} should be \code{1:v^2},
and \code{intersect(C[[i]][[j]])} for class \code{i} over all \code{j} communities should be empty.
Defaults to \code{list(list(seq(1, floor(n/2)), seq(floor(n/2)+1, n)), list(seq(1, floor(n/2)),seq(floor(n/2)+1, n)))}.}

\item{P}{the matrix of probabilities, where \code{P[[i]][k]} corresponds to the probability of an edge for class \code{i} in edge community \code{k}.
Defaults to \code{list(c(0.5, 0.5), c(0.5, 0.5))}.}
}
\value{
An object of class \code{"Simulation"} containing the following:
\item{\code{graphs}}{\code{[[n]][v, v]} the \code{n} graphs with \code{v} vertices as a list of adjacency matrices.}
\item{\code{Y}}{\code{[n]} the class labels for each of the \code{n} graphs.}
}
\description{
A function to simulate graphs from the Structured Independent Edge Model (SIEM).
}
\examples{
library(graphstats)
data <- gs.sims.siem(100, 10)  # simulate 100 graphs with 10 vertices from the default model
}
\author{
Eric Bridgeford
}
