% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipar.R
\name{gs.test.semipar}
\alias{gs.test.semipar}
\title{Semiparametric two-sample testing}
\usage{
gs.test.semipar(g1, g2, k = NULL, nsamples = 2 * max(gorder(g1),
  gorder(g2)), test = "R", verbose = FALSE)
}
\arguments{
\item{g1}{input graph, as an igraph object. See \code{\link[igraph]{graph}} for details.}

\item{g2}{input graph, as an igraph object. See \code{\link[igraph]{graph}} for details.}

\item{k}{dimension of the latent position that graphs are embeded into.  Defaults to \code{NULL} which selects the optimal dimensionality using \link{gs.dim.select}.}

\item{nsamples}{Number of bootstrap samples when performing hypothesis tesing. Defaults to \code{2*n} where \code{n} is the number of vertices, \code{max(gorder(g1), gorder(g2))}.}

\item{test}{the type of test statistic to use. Defaults to \code{'R'}. Supported options are:
\itemize{
\item{\code{'R'}}{estimate only a rotation between the latent positions of \code{g1} and \code{g2} when computing the test statistic.}
}}

\item{verbose}{logical indicating whether to print output to console. Defaults to \code{FALSE}.}
}
\value{
a list containing the following:
\item{\code{T.obs}}{The observed test statistic for whether \code{g1} and \code{g2} are sampled from the same distribution. If \code{T.obs} is near 0, this indicates that \code{g1} and \code{g2} likely come from the same distribution.}
\item{\code{T.null}}{A length \code{nsamples} vector indicating the test statistic under the bootstrapped samples for whether \code{g1} and \code{g2} are sampled from the same distribution.}
\item{\code{p.value}}{the p-value associated with the semiparametric two-sample test; the fraction of times \code{T.alt < T.obs}.}
}
\description{
This is a simple implementation of the semiparametric
two-sample testing problem of given \eqn{X_1, X_2, \dots, X_n} i.i.d. \eqn{F} and
\eqn{Y_1, Y_2, \dots, Y_m} i.i.d. \eqn{G}, test the null hypothesis of \eqn{F = G} against
the alternative hypothesis of \eqn{F \not = G}.
}
\references{
Athreya, A., Fishkind, D. E., Levin, K., Lyzinski, V., Park, Y., Qin, Y., Sussman, D. L., Tang, M., Vogelstein, J. T., Priebe, C. E.
Statistical inference on random dot product graphs: a survey
}
\author{
Eric Bridgeford <ericwb95@gmail.com>
}
