% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gs.utils.check_params}
\alias{gs.utils.check_params}
\title{Stochastic-Block Model Graphs}
\usage{
gs.utils.check_params(params)
}
\arguments{
\item{params}{a named list of the model parameters. Assumed this list contains some element named \code{params$priors}.}
}
\value{
K the number of classes to simulate graphs from.
}
\description{
A function to check whether all parameters specified for a simulation are appropriately sized.
}
\author{
Eric Bridgeford
}
