% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{gs.xfm.aaply}
\alias{gs.xfm.aaply}
\title{Apply Over array of graphs and return array of graphs}
\usage{
gs.xfm.aaply(Xs, func, rparam = NULL, ...)
}
\arguments{
\item{Xs}{[n, v, v] the set of n graphs with v vertices in an array.}

\item{func}{the function to apply to each of the n graphs. Should take as its first argument a [v, v] matrix, and return ONLY a [v, v] matrix.}

\item{...}{the additional arguments that would otherwise be passed to func, ordered or keyworded as necessary.}

\item{rparam=NULL}{if func does not return a [v, v] matrix but has one particular parameter that is, specify it here.}
}
\value{
Rs [n, v, v] the set of n graphs wit v vertices, transformed according to func
}
\description{
A function to streamline applying a function over a set of graphs as an array and producing an array.
}
\author{
Eric Bridgeford
}
