% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{gs.xfm.alply}
\alias{gs.xfm.alply}
\title{Apply Over array of graphs and return list per graph}
\usage{
gs.xfm.alply(Xs, func, ...)
}
\arguments{
\item{Xs}{[n, v, v] the set of n graphs with v vertices in an array.}

\item{func}{the function to apply to each of the n graphs. Should take as its first argument a [v, v] matrix.}

\item{...}{the additional arguments that would otherwise be passed to func, ordered or keyworded as necessary.}
}
\value{
Rs [[n]] the returns of func applied to each [v, v] entry of Xs.
}
\description{
A function to streamline applying a function over a set of graphs as an array and a list of function fits to each graph.
}
\author{
Eric Bridgeford
}
