% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{gs.xfm.rank_graph}
\alias{gs.xfm.rank_graph}
\title{Rank Graph}
\usage{
gs.xfm.rank_graph(X, ties = "average", na.last = TRUE)
}
\arguments{
\item{X}{[v, v] a graph with v vertices.}

\item{ties='average'}{the method to break ties.
\itemize{
\item{'average'}{take the average rank between the tied items.}
\item{'first'}{choose the lowest index as the tie breaker.}
\item{'last'}{choose the highest index as the tie breaker.}
\item{'random'}{randomly break ties.}
\item{'max'}{replace all tied values with the maximum rank they could occupy.}
\item{'min'}{replace all tied values with the minimum rank they could occupy.}
}}

\item{na.last=TRUE}{\itemize{
\item{TRUE}{missing values in the data are put last}
\item{FALSE}{missing values in the data are put first}
}}
}
\value{
R [v, v] the rank of each edge, from lowest (1) to highest (n^2, or otherwise depending on ties).
}
\description{
A function to rank the edges of a graph.
}
\seealso{
\code{\link{stats::rank}}
}
\author{
Eric Bridgeford
}
