% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{gs.xfm.thresh_graph}
\alias{gs.xfm.thresh_graph}
\title{Threshold Graph}
\usage{
gs.xfm.thresh_graph(graph, thresh = 0.5, method = "quantile", ...)
}
\arguments{
\item{graph}{\code{[v, v]} a graph with v vertices.}

\item{thresh}{the threshold to use for binarization.}

\item{method}{the method to threshold. Defaults to \code{'quantile'}.
\itemize{
\item{'quantile'}{use the \code{thresh*100} percentile of edge values to threshold, setting values under this to 0 and over to 1.}
\item{'abs'}{use the threshold value provided explicitly to binarize, setting values under thresh to 0 and over to 1.}
}}

\item{...}{trailing args.}
}
\value{
R \code{[v, v]} the rank of each edge, from lowest (\code{1}) to highest (\code{v^2}, or otherwise depending on ties).
}
\description{
A function to threshold and binarize the edges of a graph.
}
\author{
Eric Bridgeford
}
