% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.dmri}
\alias{human.dmri}
\title{Human dMRI Connectome}
\format{An \code{\link[igraph]{igraph}} object containing the average dMRI graph.
\describe{
\item{\code{V(human.dmri)$name}}{The id in the Desikan-Killany parcellation for the vertex.}
\item{\code{V(human.dmri)$lobe}}{The lobe of the brain the parcel overlaps maximally with. \code{M} is midbrain, \code{P} is parietal, \code{O} is occipital, \code{T} is temporal, \code{F} is frontal.}
\item{\code{V(human.dmri)$hemisphere}}{The hemisphere of the brain the parcel overlaps maximally with. \code{R} is right; \code{L} is left.}
\item{\code{V(human.dmri)$region}}{A composite of the \code{lobe} and \code{hemisphere} attributes.}
\item{\code{E(human.dmri)$weight}}{The average number of fibers connecting the particular two vertices.}
}}
\usage{
human.dmri
}
\description{
A graph comprised from n=3228 diffusion MRI fiber graphs, primarily from the CoRR and MRN1313 datasets. Graphs were processed with the ndmg pipeline; see references for details.
}
\references{
Kiar, G., Bridgeford, E.W., Roncal, W.G., et al. A High-Throughput Pipeline Identifies Robust Connectomes but Troublesome Variability.
\emph{bioarxiv}, 2018.
}
\keyword{datasets}
