% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.fmri}
\alias{human.fmri}
\title{Human fMRI Connectome}
\format{An \code{\link[igraph]{igraph}} object containing the average fMRI graph.
\describe{
\item{\code{V(human.fmri)$name}}{The id in the Desikan-Killany parcellation for the vertex.}
\item{\code{V(human.fmri)$lobe}}{The lobe of the brain the parcel overlaps maximally with. \code{M} is midbrain, \code{P} is parietal, \code{O} is occipital, \code{T} is temporal, \code{F} is frontal.}
\item{\code{V(human.fmri)$hemisphere}}{The hemisphere of the brain the parcel overlaps maximally with. \code{R} is right; \code{L} is left.}
\item{\code{V(human.fmri)$region}}{A composite of the \code{lobe} and \code{hemisphere} attributes.}
\item{\code{E(human.fmri)$weight}}{The average functional connectivity between the two vertices.}
}}
\usage{
human.fmri
}
\description{
A graph comprised from n=1790 functional MRI connectivity matrices, primarily from the CoRR dataset. Graphs were processed with the ndmg pipeline; see references for details.
}
\references{
Kiar, G., Bridgeford, E.W., Roncal, W.G., et al. A High-Throughput Pipeline Identifies Robust Connectomes but Troublesome Variability.
\emph{bioarxiv}, 2018.
}
\keyword{datasets}
