% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{model.var}
\alias{model.var}
\title{SIEM Model Variance}
\usage{
model.var(p, n)
}
\arguments{
\item{p}{the probability of a particular edge existing for an edge community.}

\item{n}{the number of edges for this particular edge community.}
}
\value{
var the variance associated with a particular edge community.
}
\description{
A function that computes the variance per-edge for SIEM.
}
\author{
Eric Bridgeford
}
