% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{nonpar}
\alias{nonpar}
\title{Nonparametric two-sample testing using kernel-based test statistic}
\usage{
nonpar(G1, G2, dim = NULL, sigma = NULL, alpha = 0.05,
  bootstrap_sample = 200, verbose = FALSE)
}
\arguments{
\item{dim}{dimension of the latent position that graphs are embeded into.  Defaults to \code{NULL} which selects the optimal dimensionality using \link{}.}

\item{sigma}{bandwidth of the rbf kernel for computing test statistic}

\item{alpha}{Significance level of hypothesis testing. Defaults to \code{.05}.}

\item{bootstrap_sample}{Number of bootstrap samples when performing hypothesis tesing}

\item{verbose}{logical indicating whether to print output to console. Defaults to \code{FALSE}.}

\item{g1}{an igraph object}

\item{g2}{an igraph object}
}
\value{
\code{T} A scalar value \eqn{T} such that \eqn{T} is near 0 if the rows of
\eqn{X} and \eqn{Y} are from the same distribution and \eqn{T} far from 0 if the rows of
\eqn{X} and \eqn{Y} are from different distribution.
}
\description{
This is a simple implementation of the kernel-based test statistic for the nonparametric
two-sample testing problem of given \eqn{X_1, X_2, \dots, X_n} i.i.d. \eqn{F} and
\eqn{Y_1, Y_2, \dots, Y_m} i.i.d. \eqn{G}, test the null hypothesis of \eqn{F = G} against
the alternative hypothesis of \eqn{F \not = G}. The test statistic is based on embedding
\eqn{F} and \eqn{G} into a reproducing kernel Hilbert space and then compute a distance between
the resulting embeddings. For this primitive, the Hilbert space is associated with the
Gaussian kernel.
}
\references{
Tang, M., Athreya, A., Sussman, D.L., Lyzinski, V., Priebe, C.E.
A nonparametric two-sample hypothesis testing problem for random graphs
}
\author{
Eric Bridgeford \href{mailto:ericwb95@gmail.com}{ericwb95@gmail.com}, Youngser Park \href{mailto:youngser@jhu.edu}{youngser@jhu.edu}, Kemeng Zhang \href{mailto:kzhang@jhu.edu}{kzhang@jhu.edu}.
}
