% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{ptr}
\alias{ptr}
\title{Run pass-to-rank on a weighted graph.}
\usage{
ptr(g)
}
\arguments{
\item{g}{a graph in \code{igraph} format or an n x 2 edge list or an n x n adjacency matrix}
}
\description{
It extracts (non-zero) edge weight vector \eqn{W} from a graph and replaces it with \eqn{2*R / (|E|+1)} where \eqn{R} is the rank of \eqn{W} and \eqn{|E|} is the number of edges. This does 'no-op' for an unweighted graph.
}
\author{
Youngser Park \href{mailto:youngser@jhu.edu}{youngser@jhu.edu}
}
