% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.R
\name{rsp}
\alias{rsp}
\title{Find a suitable initial position for multistart SGM}
\usage{
rsp(n, gamma)
}
\arguments{
\item{n}{the dimension to make the output matrix}

\item{gamma}{a number between 0 and 1 input
by the user as an upper bound for step \eqn{alpha}}
}
\value{
a random doubly stochastic matrix
}
\description{
Generates a random matrix to use as a starting point of the
multistart SGM algorithm. For some \eqn{alpha} in (0,\code{gamma}),
output matrix is weighted towards the barycenter by \eqn{(1-alpha)}
and towards another randomly generated doubly-stochastic matrix by
\eqn{alpha}.
}
\author{
Heather Gaddy Patsolic <hgaddy1@jhu.edu>
}
