% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_subgraph.R
\name{sg.bern.compute_graph_statistics}
\alias{sg.bern.compute_graph_statistics}
\title{Bernoulli Subgraph Computation}
\usage{
sg.bern.compute_graph_statistics(samp, Y, e, coherent = FALSE,
  tstat = "fisher")
}
\arguments{
\item{samp}{a list or array of graphs with arbitrary labelling.
 - if samp is a list, then it should have s elements of dimensions
     [n x n].
- if samp is an array, then it should be of dimensions [n x n x s].}

\item{Y}{[s] the class labels.}

\item{e}{the number of edges for the subgraph.}

\item{coherent=FALSE}{a logical indicating whether to approximate a coherent, or incoherent, subgraph.}

\item{tstat="fisher"}{the test statistic to use. options are fisher's exact ("fisher") and chi-squared ("chisq").}
}
\value{
subgraph [n x n] an array indicating whether an edge is present or not present in the subgraph.

p [n x n x c] the probability per edge of being connected per class for c classes.
}
\description{
\code{sg.bern.compute_subgraph} estimates the edges for a subgraph in a given set of samples.
}
\seealso{
\code{\link{sg.bern.graph_estimator}}  \code{\link{sg.bern.subgraph_estimator}} \code{\link{sg.bern.subgraph_estimator}}
}
