% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_subgraph.R
\name{sg.bern.estimator}
\alias{sg.bern.estimator}
\title{A function to fit a bernoulli distribution to a given univariate samp of data.}
\usage{
sg.bern.estimator(samp, thresh = 0, smooth = TRUE)
}
\arguments{
\item{samp}{a univariate samp of values with arbitrary class labels.}

\item{thresh=0:}{is the threshold below which we set edges to disconnected, and above which we set edges to connected.}

\item{smooth=TRUE:}{whether to smooth p to avoid undesirable limits.}
}
\value{
p the p parameter per edge.
}
\description{
\code{sg.bern.estimator} estimate the parameter of a bernoulli distribution, p, for a given samp.
}
\seealso{
\code{\link{sg.bern.graph_estimator}}
}
