% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_subgraph.R
\name{sg.bern.sample_graph}
\alias{sg.bern.sample_graph}
\title{Sample Graphs from Bernoulli Distribution}
\usage{
sg.bern.sample_graph(p, s = 10, type = "array", rewire = NaN,
  directed = FALSE)
}
\arguments{
\item{p}{a [n x n] matrix indicating the alphas of each edge.}

\item{type}{an option of "list" or "array" (default) for the format of the output.}

\item{rewire}{the probability of arbitrarily rewiring edges.}

\item{n}{the number of graphs to sample.}
}
\value{
sample the graph observations from the given beta distribution.
    - if type == "list", returns a p element list of [n x m] observations.
    - if type == "array" (default), returns a [n x m x s] element array where the 3rd dimension indexes the observations.
}
\description{
\code{sg.bern.sample_graph} A function to sample from a graph-valued RV where each edge follows a bernoulli distribution.
}
\seealso{
\code{\link{list2array}} \code{\link{array2list}}
}
