% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_subgraph.R
\name{sg.bern.subgraph_classifier}
\alias{sg.bern.subgraph_classifier}
\title{Bernoulli Subgraph Classifier}
\usage{
sg.bern.subgraph_classifier(test, edges, p, pi, classes)
}
\arguments{
\item{test}{the test array to generate labels for.
\itemize{
\item if test is an array, then it should be of dimensions \link{n x n x s}.
}}

\item{p}{\link{n x n x c} the probability per edge of being connected per class for c classes.}

\item{pi}{\link{c} the pi vector probability of each class occuring.}

\item{classes}{the labels for the classes as ordered in the pi and p arrays.}
}
\value{
edges \link{e} the edges present in the subgraph.

Yhat \link{s} the predicted class for each test example.
}
\description{
\code{sg.bern.subgraph_classifier} classifies arbitrary data given the estimators of a subgraph.
}
\seealso{
\code{\link{sg.bern.compute_graph_statistics}}
}
