% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_subgraph.R
\name{sg.bern.subgraph_edge_estimation}
\alias{sg.bern.subgraph_edge_estimation}
\title{Bernoulli Subgraph Edge Estimation}
\usage{
sg.bern.subgraph_edge_estimation(tstats, e, coherent = FALSE)
}
\arguments{
\item{tstats}{\link{n x n} the test-statistic results.}

\item{e}{the number of edges to look for, arbitrarily breaking ties as necessary.}

\item{coherent=FALSE}{if FALSE, estimate an incoherent subgraph, otherwise an integer indicating the number of vertices in the coherent subgraph.}
}
\value{
edges \link{n*n} the edges in the graph ordered by test statistic depending on the estimator type.
}
\description{
\code{sg.bern.subgraph_classifier} orders the edges by test statistic results depending on the estimator type specified.
}
\seealso{
\code{\link{sg.bern.compute_graph_statistics}}
}
