% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_subgraph.R
\name{sg.bern.subgraph_estimator}
\alias{sg.bern.subgraph_estimator}
\title{Bernoulli Subgraph Estimator}
\usage{
sg.bern.subgraph_estimator(samp, Y)
}
\arguments{
\item{samp}{[n x n x s] an array of graph samples. Should be binary (0 or 1).}

\item{Y}{[s] the class labels.}
}
\value{
cont_matrix [n, n, c, 2] a contingency table for each edge, where each contingency
        table has either connected or unconnected for each class.
}
\description{
\code{sg.beta.subgraph_estimator} A function to produce estimators for bernoulli graphs needed to compute subgraphs.
}
\seealso{
\code{\link{sg.bern.graph_estimator}}
}
