% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_subgraph.R
\name{sg.bern.xval_classifier}
\alias{sg.bern.xval_classifier}
\title{Cross-Validated Bernoulli Subgraph Classifier}
\usage{
sg.bern.xval_classifier(samp, Y, nedges, coherent = FALSE, tstat = "fisher",
  xval = "loo", folds = NaN)
}
\arguments{
\item{samp}{a list or array of graphs with arbitrary labelling.
- if samp is a list, then it should have s elements of dimensions
\link{n x n}.
\itemize{
\item if samp is an array, then it should be of dimensions \link{n x n x s}.
}}

\item{Y}{\link{s} the class labels.}

\item{nedges}{\link{z} an array where each element is the number of edges to look for, arbitrarily breaking ties as necessary.}

\item{coherent=FALSE}{if FALSE, estimate an incoherent subgraph, otherwise an integer indicating the number of vertices in the coherent subgraph.}

\item{tstat="fisher"}{the test statistic to use. options are fisher's exact ("fisher") and chi-squared ("chisq").}

\item{xval="loo"}{the cross-validation options to use. Options are "loo" (leave-one-out) and "kfold" (K-fold).}

\item{folds=NaN}{the number of folds to do if xval is set to kfold.}
}
\value{
subgraph \link{n x n} an array indicating whether an edge is present or not present in the subgraph.

p \link{n x n x c} the probability per edge of being connected per class for c classes.
}
\description{
\code{sg.bern.xval_classifier} Bernoulli Subgraph Classifier with Cross Validation to determine the optimal subgraph.
}
\seealso{
\code{\link{sg.bern.compute_graph_statistics}}
}
