% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localnbd.R
\name{vnsgm.ordered}
\alias{vnsgm.ordered}
\title{Nominates vertices in second graph to match to VOI in first graph --
first reducing the problem, structurally, using provided seeds}
\usage{
vnsgm.ordered(x, S, g1, g2, h, ell, R, g, pad = 0, sim = TRUE,
  verb = FALSE, plotF = FALSE)
}
\description{
Nominates vertices in second graph to match to VOI in first graph --
first reducing the problem, structurally, using provided seeds
}
