% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\name{gs.embed.ase}
\alias{gs.embed.ase}
\title{Spectral Embedding of Adjacency Matrices}
\usage{
gs.embed.ase(g, k, edge.attr = NULL)
}
\arguments{
\item{g}{The input graph, directed or undirected.}

\item{k}{An integer scalar. Should be the case that \code{k < gorder(g)}. The
largest \code{k}-dimensions are retained from the spectral embedding.}

\item{edge.attr}{the names of the attribute to use for weights. Should be in `names(get.edge.attribute(graph))`. Defaults to \code{NULL}, which assumes the graph is binary.}
}
\value{
A list containing the following:
\item{\code{X}}{an \code{n} by \code{k} matrix indicating the estimated latent positions, where \code{n} is the number of vertices of \code{g}.}
\item{\code{Y}}{\code{NULL} if \code{g} is undirected. If \code{g} is directed, \code{Y} is a \code{n} by \code{k} matrix indicating the second half of the latent positions.}
\item{D}{The eigenvalues (for undirected graphs) or the singular values (for directed graphs) associated with the latent positions.}
\item{options}{A named list, information about the underlying ARPACK computation. See \code{\link{arpack}} for the details.}
}
\description{
Spectral decomposition of the adjacency matrices of graphs.
}
\details{
This function computes a \code{k}-dimensional Euclidean representation of
the graph based on its adjacency matrix, \eqn{A}. This representation is
computed via the singular value decomposition of the adjacency matrix,
\eqn{A=UDV^T}. In the case, where the graph is a random dot product graph
generated using latent position vectors in \eqn{R^{k}} for each vertex, the
embedding will provide an estimate of these latent vectors.

For undirected graphs the latent positions are calculated as
\eqn{X=U^{k}D^{1/2}}{U[k] sqrt(D[k])}, where \eqn{U^{k}}{U[k]} equals
to the first \code{k} columns of \eqn{U}, and \eqn{D^{1/2}}{sqrt(D[k])} is
a diagonal matrix containing the top \code{k} singular values on the
diagonal.

For directed graphs the embedding is defined as the pair
\eqn{X=U^{k}D^{1/2}}{U[k] sqrt(D[k])} and \eqn{Y=V^{k}D^{1/2}}{V[k]
sqrt(D[k])}. (For undirected graphs \eqn{U=V}, so it is enough to keep one
of them.)
}
\examples{

## A small graph
lpvs <- matrix(rnorm(200), 20, 10)
lpvs <- apply(lpvs, 2, function(x) { return (abs(x)/sqrt(sum(x^2))) })
RDP <- sample_dot_product(lpvs)
embed <- embed_adjacency_matrix(RDP, 5)
}
\references{
Sussman, D.L., Tang, M., Fishkind, D.E., Priebe, C.E.  A
Consistent Adjacency Spectral Embedding for Stochastic Blockmodel Graphs,
\emph{Journal of the American Statistical Association}, Vol. 107(499), 2012
}
\seealso{
\code{\link{sample_dot_product}}
}
\author{
Gabor Csardi \email{csardi.gabor@gmail.com}
}
\keyword{graphs}
