% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gs.plot.plot_matrix}
\alias{gs.plot.plot_matrix}
\title{Graph plot}
\usage{
gs.plot.plot_matrix(mtx, title = "", xlabel = "", ylabel = "",
  legend.name = "metric", legend.show = TRUE, font.size = 12,
  limits = c(min(mtx), max(mtx)), vfactor = FALSE, ffactor = FALSE,
  vlist = NULL)
}
\arguments{
\item{title}{the title for the square plot. Defaults to \code{""}.}

\item{xlabel}{the x label for the square plot. Defaults to \code{""}.}

\item{ylabel}{the y label for the square plot. Defaults to \code{""}.}

\item{legend.name}{the legend title for the square plot. Defaults to \code{"metric"}.}

\item{legend.show}{whether to show the legend on the plot. Defaults to \code{TRUE}.}

\item{font.size}{the default font size for the plot text. Axis/legend text is \code{font.size - 2}. Defaults to \code{12}.}

\item{limits}{the limits for fill color. Defaults to \code{c(min(graph), max(graph))}.}

\item{vfactor}{if TRUE, interpret the vertex labels as a factor. Defaults to \code{FALSE}.}

\item{ffactor}{if TRUE, interpret the fill colors as a factor. Defaults to \code{FALSE}.}

\item{vlist}{an optional list for the vertex labels ordered according to the adjacency matrix \code{graph}. Defaults to \code{NULL}.}

\item{graph}{\code{[v, v]} the input graph.}
}
\value{
plot the matrix as a plot.
}
\description{
A function that plots a graph.
}
\author{
Eric Bridgeford
}
