% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{gs.sbm.fit}
\alias{gs.sbm.fit}
\title{Stochastic Block Model}
\usage{
gs.sbm.fit(g, community.attribute)
}
\arguments{
\item{g}{an \code{igraph} object. See \code{\link[igraph]{graph}} for details.}
}
\value{
the probability matrix for each block, as an \code{igraph} object.
Vertices will be the unique communities for attribute \code{community.attribute}. Each vertex in the resulting graph will have an attribute \code{community} indicating the vertices in \code{graph} that
comprise the particular community the vertex summarizes See \code{\link[igraph]{graph}} for details.
}
\description{
A function to fit a stochastic block model to a weighted or unweighted graph.
}
\author{
Eric Bridgeford
}
