% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.batch.tstat}
\alias{gs.siem.batch.tstat}
\title{Test Statistic for Batch Detection}
\usage{
gs.siem.batch.tstat(models, Z, i = 1, j = 2)
}
\arguments{
\item{models}{[[n]] a list of the models fit to each of the n samples.}

\item{Z}{[n] an array of the labels associated with each model.}

\item{i=1}{the first edge set to use in the comparison.}

\item{j=2}{the second edge set to use in the comparison.}
}
\value{
tstat the test statistic.
}
\description{
A function that computes a test statistic associated with a set of estimators.
}
\author{
Eric Bridgeford
}
