% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siem.R
\name{gs.siem.fit}
\alias{gs.siem.fit}
\title{Structured Independent Edge Model, Single Sample}
\usage{
gs.siem.fit(X, Es, alt = "greater", i = NULL, j = NULL)
}
\arguments{
\item{X}{[v, v] the binary graph with v vertices.}

\item{Es}{[[e]] a numbered list where each element corresponds to a community of edges. Each community should consist of the 1-dimensional indices of an edge.}

\item{alt='greater'}{the alternative hypothesis for each edge test.
\itemize{
\item{'greater'}{p corresponds to H0: x1 <= x2, HA: x1 > x2. R(x1, x2) = x1 > x2}
\item{'neq'}{p corresponds to H0: x1 = x2, HA: x1 != x2. R(x1, x2) = x1 != x2}
\item{'less'}{p corresponds to H0: x1 >= x2, HA: x1 < x2. R(x1, x2) = x1 < x2}
}}
}
\value{
pr [e] a probability array corresponding to the probability of an edge connecting each edge community.

var [e] the variance of the probabilities estimated between edge communities.

dpr [e, e] an array consisting of the paired differences in probability where dpr_{ij} = pr_i - pr_j.

dvar [e, e] an array consisting of the variance of the estimate of the paired differences in probability, where dvar_{ij} = var_i + var_j.

pv [e, e] pv_{ij} is the p-value of false rejection of H0 that !R(pr_i, pr_j) in favor of HA that R(pr_j, pr_i).
}
\description{
An independent-edge generalization of the traditional Stochastic Block Model.
}
\author{
Eric Bridgeford
}
