% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{gs.xfm.rank_graph}
\alias{gs.xfm.rank_graph}
\title{Rank Graph}
\usage{
gs.xfm.rank_graph(graph, ties = "average", na.last = TRUE, ...)
}
\arguments{
\item{graph}{\code{[v, v]} a graph with \code{v} vertices.}

\item{ties}{the method to break ties. Defaults to \code{'average'}.
\itemize{
\item{'average'}{take the average rank between the tied items.}
\item{'first'}{choose the lowest index as the tie breaker.}
\item{'last'}{choose the highest index as the tie breaker.}
\item{'random'}{randomly break ties.}
\item{'max'}{replace all tied values with the maximum rank they could occupy.}
\item{'min'}{replace all tied values with the minimum rank they could occupy.}
}}

\item{na.last}{How to interpret missing values in the data. Defaults to \code{TRUE}.
\itemize{
\item{TRUE}{missing values in the data are put last}
\item{FALSE}{missing values in the data are put first}
}}

\item{...}{trailing args.}
}
\value{
R \code{[v, v]} the rank of each edge, from lowest (1) to highest (n^2, or otherwise depending on ties).
}
\description{
A function to rank the edges of a graph.
}
\seealso{
\code{\link{stats::rank}}
}
\author{
Eric Bridgeford
}
