% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{gs.xval.split}
\alias{gs.xval.split}
\title{Cross-Validation Data Splitter}
\usage{
gs.xval.split(samp, Y, k = "loo")
}
\arguments{
\item{samp}{[v, v, s] an array of input data.}

\item{Y}{[s] the class labels.}

\item{k='loo'}{the cross-validated method to perform.
\itemize{
\item{'loo'}{Leave-one-out cross validation}
\item{isinteger(k)}{perform k-fold cross-validation with k as the number of folds.}
}}
}
\value{
sets the cross-validation sets.
}
\description{
\code{sg.bern.xval_split_data} A function to split a series of graphs into
training and testing sets for cross validation.
}
