% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiStart.r
\name{multiStart}
\alias{multiStart}
\title{Multiple Restart SGM}
\usage{
multiStart(A, B, R, s, gamma, pad = 0, iter = 20)
}
\arguments{
\item{A}{adjacency matrix for \eqn{G_1}}

\item{B}{adjacency for correlated matrix \eqn{G_2}}

\item{R}{number of restarts}

\item{s}{number of seeds; assumed to be first \eqn{s} vertices in both
\eqn{G_1} and \eqn{G_2}}

\item{gamma}{(upper bound on alpha) should be in (0,1), max tolerance
for alpha, how far away from the barycenter user is willing to go for
the initialization of \code{sgm} on any given iteration}

\item{pad}{a scalar value for padding for sgm (cases where two graphs
have different number of vertices); defaulted to 0}

\item{iter}{number of iterations allowed to be used in sgm step}
}
\value{
\eqn{n \times n} average probability matrix over all restarts.
}
\description{
Runs SGM algorithm with random starting point generated by \code{rsp} \code{R}
times and generates an average probability matrix for matched vertex
over the \code{R} restarts
}
\author{
Heather Gaddy Patsolic <hgaddy1@jhu.edu>
}
