% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnsgm.R
\name{vnsgm}
\alias{vnsgm}
\title{Vertex Nomination via Seeded Graph Matching}
\usage{
vnsgm(x, seeds, g1, g2, h, ell, R, gamma, maxiter = 20, pad = 0,
  verbose = FALSE, plotF = FALSE)
}
\arguments{
\item{x}{vector of indices for vertices of interest (voi) in \eqn{G_1}}

\item{seeds}{a numeric matrix, the \eqn{m \times 2}{m*2} matching vertex table,
where m is the number of seeds.}

\item{g1}{\eqn{G_1} in \code{igraph} object where voi is known}

\item{g2}{\eqn{G_2} in \code{igraph}}

\item{h}{\eqn{h}-hop for distance from voi to other vertices to create
\eqn{h}-hop induced subgraph of \eqn{G_1}}

\item{ell}{\eqn{ell}-hop for distance from seeds to other vertices
to create \eqn{ell}-hop induced subgraph of \eqn{G_1}}

\item{R}{number of restarts for \code{multiStart}}

\item{gamma}{to be used with \code{multiStart}, max tolerance
for alpha, how far away from the barycenter user is willing to go for
the initialization of \code{sgm} on any given iteration}

\item{maxiter}{the number of maxiters for the Frank-Wolfe algorithm.}

\item{pad}{a scalar value for padding for sgm (cases where two graphs
have different number of vertices); defaulted to 0}

\item{verbose}{logical verbose outputs}

\item{plotF}{boolean to plot the probability matrix}
}
\value{
\code{VOI}  Vertex of Interest

\code{seeds} \code{s} seeds

\code{Sx1} \eqn{S_x} the seeds within an h-path, i.e. in the
h-neighborhood, of VOI x in \eqn{G_1}

\code{Sx2} \eqn{S'_x} the corresponding seeds of \eqn{S_x} in \eqn{G_2}

\code{candidates_for_matching} labels for the candidates for
matching VOIs in \eqn{G_2}

\code{G1_vertices} vertices within ell-neighborhood of S_x in \eqn{G_1}
(vertices used in vertex nomination for VOIs)

\code{G2_vertices} vertices within ell-neighborhood of S'_x in \eqn{G_2}
(including candidates x' for matching VOIs x in \eqn{G_1})

\code{P} matrix \eqn{P(i,j)} is the proportion of times that vertex \eqn{j} in
the induced subgraph of \eqn{G_2} was mapped to vertex \eqn{i} in the induced subgraph of \eqn{G_1}.
Then the \eqn{i-th} and \eqn{j-th} elements of the labels vector tells you which vertices these actually were
in \eqn{G_1} and \eqn{G_2}, respectively.
}
\description{
Finds the seeds in an \eqn{h}-hop induced nbd of \eqn{G_1} around the
VOI, x,that is, finds induced subgraph generated by vertices that are
within a path of length \eqn{h} the VOI, and then finds an \eqn{ell}-hop
induced nbd of \eqn{G_1} around the seeds within the \eqn{h}-hop nbd of
x, and an \eqn{ell}-hop induced nbd of \eqn{G_2} around the corresponding
seeds. Then, matches these induced subgraphs via \code{multiStart}.
}
\references{
Patsolic, Heather G.; Park, Youngser; Lyzinski, Vince; Priebe, Carey E. (2017).
Vertex Nomination Via Local Neighborhood Matching
Online: \url{https://arxiv.org/abs/1705.00674}

Fishkind, D. E., Adali, S., Priebe, C. E. (2012). Seeded Graph Matching
Online: \url{http://arxiv.org/abs/1209.0367}
}
\author{
Youngser Park \href{mailto:youngser@jhu.edu}{youngser@jhu.edu}, Kemeng Zhang \href{mailto:kzhang@jhu.edu}{kzhang@jhu.edu}
}
